requirements.txt
pandas==2.2.2
numpy==2.0.1
requests==2.32.3
python-dotenv==1.0.1
pytrends==4.9.2
matplotlib==3.9.0
plotly==5.23.0
jupyterlab==4.2.4
pydantic==2.8.2
pyyaml==6.0.2
tqdm==4.66.4
